requirements.txt
numpy==1.24.2
torch==2.0.1
torchvision==0.15.2
pillow==9.4.0
opencv-python==4.7.0.72
scikit-image==0.20.0
scipy==1.10.1
# yaml==0.2.5
pyyaml==6.0
easydict==1.10
replicate==0.5.1
django==4.1.7
djangorestframework==3.14.0
dlib==19.24.0
flask==2.2.3
dill==0.3.6
tensorboardx==2.6
python-dotenv==1.0.0
django-cors-headers
matplotlib==3.7.1